+++ Source/EventFlow.Tests/IntegrationTests/InMemoryConfiguration.cs
@@ -24,6 +24,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Configuration;
+using EventFlow.Core;
 using EventFlow.Extensions;
 using EventFlow.Queries;
 using EventFlow.ReadStores;