+++ Source/EventFlow/Aggregates/IAggregateEvent.cs
@@ -20,6 +20,8 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using EventFlow.Core;
+
 namespace EventFlow.Aggregates
 {
     public interface IAggregateEvent