+++ Source/EventFlow/Aggregates/IAggregateRoot.cs
@@ -23,20 +23,31 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using EventFlow.Core;
 using EventFlow.EventStores;
 
 namespace EventFlow.Aggregates
 {
-    public interface IAggregateRoot<out TIdentity>
+    public interface IAggregateRoot
+    {
+        IAggregateName Name { get; }
+        int Version { get; }
+        IEnumerable<IAggregateEvent> UncommittedEvents { get; }
+        bool IsNew { get; }
+
+        Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(
+            IEventStore eventStore,
+            ISourceId sourceId,
+            CancellationToken cancellationToken);
+
+        void ApplyEvents(IEnumerable<IAggregateEvent> aggregateEvents);
+
+        void ApplyEvents(IReadOnlyCollection<IDomainEvent> domainEvents);
+    }
+
+    public interface IAggregateRoot<out TIdentity> : IAggregateRoot
         where TIdentity : IIdentity
     {
         TIdentity Id { get; }
-        int Version { get; }
-        bool IsNew { get; }
-        IEnumerable<IAggregateEvent> UncommittedEvents { get; }
-
-        Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(IEventStore eventStore, CancellationToken cancellationToken);
-        void ApplyEvents(IEnumerable<IAggregateEvent> aggregateEvents);
-        void ApplyEvents(IReadOnlyCollection<IDomainEvent> domainEvents);
     }
 }