+++ Source/EventFlow/Aggregates/MetadataKeys.cs
@@ -24,6 +24,7 @@ namespace EventFlow.Aggregates
 {
     public sealed class MetadataKeys
     {
+        public const string EventId = "event_id";
         public const string BatchId = "batch_id";
         public const string EventName = "event_name";
         public const string EventVersion = "event_version";
@@ -32,5 +33,6 @@ namespace EventFlow.Aggregates
         public const string AggregateSequenceNumber = "aggregate_sequence_number";
         public const string AggregateName = "aggregate_name";
         public const string AggregateId = "aggregate_id";
+        public const string SourceId = "source_id";
     }
 }