+++ Source/EventFlow/Exceptions/DomainError.cs
@@ -27,12 +27,22 @@ namespace EventFlow.Exceptions
 {
     public class DomainError : Exception
     {
-        private DomainError(string message) : base(message) { }
+        protected DomainError(string message)
+            : base(message) { }
+
+        protected DomainError(string message, Exception innerException)
+            : base(message, innerException) { }
 
         [StringFormatMethod("format")]
         public static DomainError With(string format, params object[] args)
         {
             return new DomainError(string.Format(format, args));
+        }
+
+        [StringFormatMethod("format")]
+        public static DomainError With(Exception innerException, string format, params object[] args)
+        {
+            return new DomainError(string.Format(format, args), innerException);
         }
     }
 }