+++ Source/EventFlow/ValueObjects/ValueObject.cs
@@ -72,11 +72,14 @@ namespace EventFlow.ValueObjects
             return GetProperties().Select(x => x.GetValue(this));
         }
 
-        private IEnumerable<PropertyInfo> GetProperties()
+        protected virtual IEnumerable<PropertyInfo> GetProperties()
         {
             return TypeProperties.GetOrAdd(
                 GetType(),
-                t => t.GetProperties(BindingFlags.Instance | BindingFlags.Public).OrderBy(p => p.Name).ToList());
+                t => t
+                    .GetProperties(BindingFlags.Instance | BindingFlags.Public)
+                    .OrderBy(p => p.Name)
+                    .ToList());
         }
     }
 }